GH-51 - moved search stuff to custom hook
